Start introducing Etablissement for TIAC
- Share most of the model with the existing etablissement for SSA
- Share most of the form logic with SSA
- Moving a field that is used multiple times into core
- Add a factory for new model
