+++ core/forms.py
@@ -5,10 +5,11 @@ from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.core.exceptions import ValidationError
 from django.utils.safestring import mark_safe
+from django_countries.fields import CountryField
 
 from core.form_mixins import DSFRForm, WithNextUrlMixin, WithContentTypeMixin
-from core.fields import DSFRRadioButton, ContactModelMultipleChoiceField, SEVESChoiceField
-from core.models import Document, Contact, Message, Visibilite, Structure
+from core.fields import DSFRRadioButton, ContactModelMultipleChoiceField, SEVESChoiceField, AdresseLieuDitField
+from core.models import Document, Contact, Message, Visibilite, Structure, Departement
 from core.validators import MAX_UPLOAD_SIZE_BYTES, MAX_UPLOAD_SIZE_MEGABYTES
 from core.widgets import RestrictedFileWidget
 
@@ -328,3 +329,30 @@ class AgentAddForm(DSFRForm):
         if obj:
             queryset = queryset.exclude(id__in=obj.contacts.values_list("id", flat=True))
         self.fields["contacts_agents"].queryset = queryset
+
+
+class BaseEtablissementForm(forms.ModelForm):
+    siret = forms.CharField(
+        required=False,
+        max_length=14,
+        widget=forms.HiddenInput,
+    )
+    code_insee = forms.CharField(widget=forms.HiddenInput(), required=False)
+    adresse_lieu_dit = AdresseLieuDitField(choices=[], required=False)
+    pays = CountryField(blank=True).formfield(widget=forms.Select(attrs={"class": "fr-select"}))
+    departement = forms.ModelChoiceField(
+        queryset=Departement.objects.order_by("numero").all(),
+        to_field_name="numero",
+        required=False,
+        label="Département",
+    )
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+        if not self.is_bound and self.instance and self.instance.pk and self.instance.adresse_lieu_dit:
+            self.fields["adresse_lieu_dit"].choices = [(self.instance.adresse_lieu_dit, self.instance.adresse_lieu_dit)]
+
+        departement_obj = self.instance.departement
+        if departement_obj:
+            self.initial["departement"] = departement_obj.numero
